common: Broke the USB descriptors out to their own header
