+++ src/platforms/common/cdcacm.h
@@ -30,12 +30,6 @@
 
 #include "usb.h"
 
-#define CDCACM_PACKET_SIZE 	64
-
-#define CDCACM_GDB_ENDPOINT	1
-#define CDCACM_UART_ENDPOINT	3
-#define TRACE_ENDPOINT			5
-
 void blackmagic_usb_init(void);
 /* Returns current usb configuration, or 0 if not configured. */
 int cdcacm_get_config(void);
